phase one - disconnect dependencies
