+++ org.mwc.debrief.satc.core/src/com/planetmayo/debrief/satc/zigdetector/IZigStorer.java
@@ -18,11 +18,9 @@ public interface IZigStorer
 	 * @param scenario
 	 * @param tStart
 	 * @param tEnd
-	 * @param sensor - optional sensor object that produced the data
 	 * @param rms - the %age error from the RMS for the whole leg
 	 */
-	void storeZig(String scenarioName, long tStart, long tEnd, Sensor sensor,
-			double rms);
+	void storeZig(String scenarioName, long tStart, long tEnd, double rms);
 	
 	public ArrayList<StraightLegForecastContribution> getSlices();
