+++ org.mwc.debrief.track_shift/src/org/mwc/debrief/track_shift/zig_detector/moving_average/TimeBasedMovingAverage.java
@@ -1,4 +1,4 @@
-package org.mwc.debrief.track_shift.zig_detector;
+package org.mwc.debrief.track_shift.zig_detector.moving_average;
 
 
 /**
